requirements.txt
# Build dependencies (security-critical)
pip>=26.0  # CVE-2025-8869 fix
setuptools>=80.10.2  # CVE-2024-6345, PYSEC-2025-49 fixes
wheel>=0.46.0

# Core Django
Django==6.0.3  # Major upgrade from 5.1.15 — requires Python 3.12+
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-crispy-forms==2.6
crispy-tailwind==1.0.3
django-cors-headers==4.9.0
django-htmx==1.27.0  # HTMX server-side integration

# Database
psycopg2-binary==2.9.11  # Updated from 2.9.10 (security & bug fixes)

# Authentication
django-allauth==65.15.0  # Updated from 65.1.0 (latest stable)
PyJWT==2.12.1  # Updated from 2.11.0 (CVE-2026-32597: crit header bypass)
cryptography>=44.0.1  # Required for JWT token verification with RSA keys (CVE-2026-26007, OpenSSL fixes)
google-auth==2.49.1  # Required for Google Wallet API authentication

# Storage
django-storages==1.14.6  # Updated from 1.14.2
azure-storage-blob==12.28.0
azure-identity==1.25.3  # Managed Identity & Azure CLI authentication

# Image processing
Pillow==12.1.1  # Updated from 11.0.0 (security fixes)
olefile==0.47
# MIME type detection - platform-specific packages
# Linux (Azure): python-magic requires libmagic1 (pre-installed on most Linux distros)
# Windows (dev): python-magic-bin bundles libmagic
python-magic==0.4.27; sys_platform != 'win32'
python-magic-bin==0.4.14; sys_platform == 'win32'

# Microsoft Graph API (for email)
msal==1.35.1  # Updated from 1.31.1 (latest stable)
msgraph-core==1.3.8

# Production Server
gunicorn==25.1.0
channels==4.2.0
channels-redis==4.2.1
uvicorn[standard]==0.34.0

# Utilities
python-dotenv==1.2.2  # Updated from 1.0.1
whitenoise==6.12.0  # Updated from 6.7.0 (security fix for autorefresh mode)
requests==2.32.5
httpx[http2]==0.28.1
qrcode[pil]==8.2  # QR code generation for Journey Gifts


# Push Notifications
py-vapid==1.9.4
pywebpush==2.3.0

# Azure Application Insights - SDK-based for exception filtering
# IMPORTANT: Set ApplicationInsightsAgent_EXTENSION_VERSION=disabled in Azure App Service
azure-monitor-opentelemetry==1.8.7

# Cookie Consent (GDPR/ePrivacy)
django-cookie-consent==1.0.0

# Internationalization - Model Field Translations
django-modeltranslation==0.20.2

# Testing
pytest==9.0.2
pytest-django==4.12.0
pytest-playwright==0.7.2
pytest-mock==3.15.1  # Provides 'mocker' fixture for mocking
playwright==1.58.0
